release 1.1

+++ components/admin/remove_catalog.php
@@ -3,16 +3,26 @@
 
     require('../db_config.php');   
 
-    //$photo = $_FILES['photo'];
-    $id = $_POST['id']; 
-    echo($_POST['id']);  
-    $sql = "SELECT `imgUrl` FROM `catalog` WHERE `id`=$id";
-    $result = mysqli_query($link, $sql);
-    $row = mysqli_fetch_array($result);
-    unlink('../../'.$row['imgUrl']);
-    
-    $sql = "DELETE FROM `catalog` WHERE `id`=$id";
-    $result = mysqli_query($link, $sql);
+    if(isset($_POST['action']) && $_POST['action'] === 'catalog') {
+        //$photo = $_FILES['photo'];
+        $id = $_POST['id']; 
+        
+        $sql = "SELECT `imgUrl` FROM `catalog` WHERE `id`=$id";
+        $result = mysqli_query($link, $sql);
+        $row = mysqli_fetch_array($result);
+        unlink('../../'.$row['imgUrl']);
+        
+        $sql = "DELETE FROM `catalog` WHERE `id`=$id";
+        $result = mysqli_query($link, $sql);
+    }
+
+    if(isset($_POST['action']) && $_POST['action'] === 'sales') {
+        //$photo = $_FILES['photo'];
+        $id = $_POST['id']; 
+
+        $sql = "DELETE FROM `sales` WHERE `id`=$id";
+        $result = mysqli_query($link, $sql);
+    }
     
     
     //header('Location: admin_info.php?change=catalog&type=changing');
